- Fixed minute hand movement for minutes producing decimal per minute   movement: 25 and 35

+++ src/CircularTimer.h
@@ -73,7 +73,7 @@ private:
     QTime m_startTime;
     QTime m_endTime;
 
-    int mMinuteHandMovementAngle;
+    double mMinuteHandMovementAngle;
     int mDuration;
 };
